migration/requirements.txt
firebase-admin
supabase
python-dotenv